this folder contains the BERT model
This is where we left off with the BERT model. You can put in a random string and have the model return a prediction (which is what Robert and the higher ups wanted).
The next step is to figure out how to implement this model into the existing file structure for the Blue Witness product. 

This is a great starting off point for the Labs 34 cohort for working with BERT and finding TRUE cases of Police use of force.